Use Markdown Editor in card description when editting
Since the card support markdown render, it'd be great if we also had markdown editor so for those who doesn't understand a lot about markdown could use this feature.

With this new feature, there also a drawbacks such as no longer saving when input are onBlur since if we still using this `blur` callback, it'd be called when user trying to click at the editor toolbar such as **Bold** button.