fix: Allow specifying an external dburl in helm
Some helm users might want to turn off the built-in postgres
installation, but the deployment template statically refers to a secret
created by the postgresql chart.

If postgres.enabled is set to false, we should ask for a dburl parameter
for an external postgres and use that in the environment instead.

The behavior if postgres.enabled == true remains unchanged.

Signed-off-by: Brian Stinson <brian@bstinson.com>