ref: Parametrize OIDC authorization response mode
Planka used a default response_mode 'fragment', which is not supported by all
OIDC providers.

Planka supports only the Authorization Code flow. The default response mode
for the authorization code flow is 'query', meaning the authorization server
appends the authorization code to the redirect URI as a query parameter.

I have added two environment variables: one to use the default response mode
from the OIDC provider, and one to customize the response mode if needed.

Using the default response mode is recommended by the OIDC specification:
"This use of this parameter is NOT RECOMMENDED when the Response Mode that
would be requested is the default mode specified for the Response Type."

To avoid any breaking changes, I kept the default value as 'fragment'. Ideally,
the environment variable should be undefined by default.