feat: Change the update strategy in case of persistence
In case planka is installed with a PV associated to the planka pod, to
persist attachments and avatars for example, the default update strategy
for the deployment will fail. The default "rolling update" strategy
would indeed try to start a new pod before stopping the old one. But as
the old pod is attached to the PV, the new pod will fail to attach and
so will stay in "container creating" mode, and the old pod will never be
stopped.

To avoid that, adding a condition on the deployment to use the
"recreate" update strategy if the persistence is enabled. This will
ensure the old pod is stopped and so detached from the PV before trying
to attach the new one.