fix: Add missing keys to chart values file
Without a username and email, you can not log into the account. While
they were being silently used in the template already, there was no
indication of it.

I chose not to populate these from the secret, as it would be a breaking
change to people that now manually specify it

Signed-off-by: Steve Váradi <varadiista.1@gmail.com>